backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
pandas==2.2.3
python-docx==1.1.2
openai>=1.0.0
python-dotenv==1.0.1
pydantic==2.9.2
aiofiles==24.1.0
markdown==3.7
beautifulsoup4==4.12.3
tabulate==0.9.0
